[Lesson 5] Реализовали возможность получения данных о погоде с интерента

+++ app/src/main/java/com/example/kotlin_lessons/model/WeatherDTO.kt
@@ -97,7 +97,7 @@ class WeatherDTO (
         val humidity: Long,
 
         @SerializedName("prec_mm")
-        val precMm: Long,
+        val precMm: Double,
 
         @SerializedName("prec_prob")
         val precProb: Long,
